Managed to increase the number of shown Fibonacci numbers to the 64 bit limit (in Fibonum.cpp)

+++ MEnu_VErsions/Menu/Fibonum.h
@@ -1,4 +1,4 @@
 #ifndef FIBONUM_H_
 #define FIBONUM_H_
-void Fibonum(int count, int presentF, int pastF1, int pastF2, int i);
+void Fibonum(int count, long long presentF, long long pastF1, long long pastF2, int i);
 #endif
